Fix class Product (price)

+++ catalog/models.py
@@ -20,7 +20,7 @@ class Product(models.Model):
     prod_desc = models.TextField(max_length=500, verbose_name='описание продукта')
     image = models.ImageField(upload_to='products/', verbose_name='изображение', **NULLABLE)
     category = models.ForeignKey(Category, null=False, on_delete=models.PROTECT)
-    price = models.IntegerField(verbose_name='цена')
+    price = models.FloatField(verbose_name='цена')
     created_at = models.DateTimeField(auto_now_add=True, editable=False, verbose_name='дата создания')
     updated_at = models.DateTimeField(auto_now=True, editable=False, verbose_name='дата изменения')
     in_stock = models.BooleanField(default=True, verbose_name='в наличии')
